backend/requirements.txt
# Django & DRF
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.21.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-dbbackup==4.0.2

# API & Services
requests==2.31.0
httpx==0.25.2

# AI/ML
yandex-cloud==0.7.0
transformers==4.36.2

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Serialization & Validation
pydantic==2.5.3
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
django-lint==2.5.0

# Pagination & filtering
django-filter==23.5
djangorestframework-simplejwt==5.3.2

# GIS (for maps)
django-extensions==3.2.3

# Utils
pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
